Rearranging the project to have Dependency Injection in one place
