+++ app/src/main/java/frogermcs/io/githubclient/GithubClientApplication.java
@@ -2,12 +2,14 @@ package frogermcs.io.githubclient;
 
 import android.app.Application;
 import android.content.Context;
-import android.support.annotation.VisibleForTesting;
 
 import com.frogermcs.dagger2metrics.Dagger2Metrics;
 
-import frogermcs.io.githubclient.data.UserComponent;
-import frogermcs.io.githubclient.data.api.UserModule;
+import frogermcs.io.githubclient.dependencyinjection.components.AppComponent;
+import frogermcs.io.githubclient.dependencyinjection.components.DaggerAppComponent;
+import frogermcs.io.githubclient.dependencyinjection.components.UserComponent;
+import frogermcs.io.githubclient.dependencyinjection.modules.AppModule;
+import frogermcs.io.githubclient.dependencyinjection.modules.UserModule;
 import frogermcs.io.githubclient.data.model.User;
 import timber.log.Timber;
