+++ app/src/main/java/frogermcs/io/githubclient/dependencyinjection/components/RepositoryDetailsActivityComponent.java
@@ -1,9 +1,9 @@
-package frogermcs.io.githubclient.ui.activity.component;
+package frogermcs.io.githubclient.dependencyinjection.components;
 
 import dagger.Subcomponent;
-import frogermcs.io.githubclient.ui.activity.ActivityScope;
+import frogermcs.io.githubclient.dependencyinjection.scopes.ActivityScope;
 import frogermcs.io.githubclient.ui.activity.RepositoryDetailsActivity;
-import frogermcs.io.githubclient.ui.activity.module.RepositoryDetailsActivityModule;
+import frogermcs.io.githubclient.dependencyinjection.modules.RepositoryDetailsActivityModule;
 
 /**
  * Created by Miroslaw Stanek on 23.04.15.
