+++ app/src/main/java/frogermcs/io/githubclient/dependencyinjection/modules/RepositoriesListActivityModule.java
@@ -1,12 +1,11 @@
-package frogermcs.io.githubclient.ui.activity.module;
+package frogermcs.io.githubclient.dependencyinjection.modules;
 
 import dagger.Module;
 import dagger.Provides;
 import frogermcs.io.githubclient.data.api.RepositoriesManager;
-import frogermcs.io.githubclient.ui.activity.ActivityScope;
+import frogermcs.io.githubclient.dependencyinjection.scopes.ActivityScope;
 import frogermcs.io.githubclient.ui.activity.RepositoriesListActivity;
 import frogermcs.io.githubclient.ui.activity.presenter.RepositoriesListActivityPresenter;
-import frogermcs.io.githubclient.utils.AnalyticsManager;
 
 /**
  * Created by Miroslaw Stanek on 23.04.15.
