+++ app/src/main/java/frogermcs/io/githubclient/dependencyinjection/modules/RepositoryDetailsActivityModule.java
@@ -1,9 +1,9 @@
-package frogermcs.io.githubclient.ui.activity.module;
+package frogermcs.io.githubclient.dependencyinjection.modules;
 
 import dagger.Module;
 import dagger.Provides;
 import frogermcs.io.githubclient.data.model.User;
-import frogermcs.io.githubclient.ui.activity.ActivityScope;
+import frogermcs.io.githubclient.dependencyinjection.scopes.ActivityScope;
 import frogermcs.io.githubclient.ui.activity.RepositoryDetailsActivity;
 import frogermcs.io.githubclient.ui.activity.presenter.RepositoryDetailsActivityPresenter;
