+++ app/src/main/java/frogermcs/io/githubclient/dependencyinjection/modules/SplashActivityModule.java
@@ -1,11 +1,10 @@
-package frogermcs.io.githubclient.ui.activity.module;
+package frogermcs.io.githubclient.dependencyinjection.modules;
 
 import dagger.Module;
 import dagger.Provides;
-import frogermcs.io.githubclient.HeavyExternalLibrary;
 import frogermcs.io.githubclient.HeavyLibraryWrapper;
 import frogermcs.io.githubclient.data.api.UserManager;
-import frogermcs.io.githubclient.ui.activity.ActivityScope;
+import frogermcs.io.githubclient.dependencyinjection.scopes.ActivityScope;
 import frogermcs.io.githubclient.ui.activity.SplashActivity;
 import frogermcs.io.githubclient.ui.activity.presenter.SplashActivityPresenter;
 import frogermcs.io.githubclient.utils.Validator;
