+++ app/src/main/java/frogermcs/io/githubclient/ui/activity/BaseActivity.java
@@ -3,9 +3,6 @@ package frogermcs.io.githubclient.ui.activity;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
-import frogermcs.io.githubclient.AppComponent;
-import frogermcs.io.githubclient.GithubClientApplication;
-
 /**
  * Created by Miroslaw Stanek on 23.04.15.
  */
