+++ app/src/main/java/frogermcs/io/githubclient/ui/activity/SplashActivity.java
@@ -2,7 +2,6 @@ package frogermcs.io.githubclient.ui.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Debug;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -19,7 +18,7 @@ import butterknife.OnClick;
 import frogermcs.io.githubclient.GithubClientApplication;
 import frogermcs.io.githubclient.R;
 import frogermcs.io.githubclient.data.model.User;
-import frogermcs.io.githubclient.ui.activity.module.SplashActivityModule;
+import frogermcs.io.githubclient.dependencyinjection.modules.SplashActivityModule;
 import frogermcs.io.githubclient.ui.activity.presenter.SplashActivityPresenter;
 import frogermcs.io.githubclient.utils.AnalyticsManager;
 import rx.Subscription;
