+++ app/src/main/java/frogermcs/io/githubclient/utils/AnalyticsManager.java
@@ -2,11 +2,6 @@ package frogermcs.io.githubclient.utils;
 
 import android.app.Application;
 
-import javax.inject.Inject;
-import javax.inject.Singleton;
-
-import frogermcs.io.githubclient.data.UserScope;
-import frogermcs.io.githubclient.ui.activity.ActivityScope;
 import timber.log.Timber;
 
 /**
